refactor component hierarchy & route handling
This commit is in preparation for supporting upcoming gallery pages.
All scaffolding/Menu code was abstracted into a generic <App>
shell-component.

Each animation now has it's own `/<animation-uuid>` path, which will
open the Editor with said animation. Before a redux state-variable was
used to store the "selectedAnimation", instead of a route parameter.