add public gallery page
A new page which displays all animations in the public gallery. The
Public gallery is a moderated set of user-generated publicly viewable
animations, which users can copy to their personal library.

Next up is adding the feature to copy an animation from the gallery to
personal library. Further will we introduce an admin view of this page,
which acts as the moderation tool in the backend.